Rename k8s package to client
